+++ tools/util.cxx
@@ -325,7 +325,7 @@ HString get_token( yaal::hcore::HString const& s, yaal::hcore::HString const& d,
 
 bool is_byte( int value )
 	{
-	return ( value <= UCHAR_MAX );
+	return ( value <= static_cast<int>( UCHAR_MAX ) );
 	}
 
 void show_help( void* arg )
